Hr permissions

+++ desktop-app/src/main/java/com/sanitas/clinicapp/hr/HrController.java
@@ -91,9 +91,12 @@ public class HrController {
             }
 
             for (int index : indexes) {
+                if(model.viewRole((String) employeeTable.getValueAt(index, 3))=="administrator" && (!account.isSuperAdmin())) {
+                    view.sendError("Nu aveti permisiunile necesare!");
+                    continue;
+                }
                 model.deleteEmployee((String) employeeTable.getValueAt(index, 3));
             }
-
             view.getViewv().updateTable(model.getAllData("", "", ""));
             view.sendSuccessMessage("Angajatii selectati au fost stersi.");
         }
@@ -113,7 +116,7 @@ public class HrController {
                 List<Schedule> schedules=model.viewSchedule(employee.getCnp());
                 PanelViewSchedule panel = new PanelViewSchedule(model,schedules,account);
                 panel.addDeleteButtonListener(new DeleteScheduleListener(employee.getCnp()));
-                panel.addInsertButtonListener(new InsertScheduleListener(employee.getCnp(),schedules));
+                panel.addInsertButtonListener(new InsertScheduleListener(employee.getCnp(), schedules));
                 panel.addUpdateButtonListener(new UpdateScheduleListener(employee.getCnp()));
                 panel.addBackButtonListener(new BackScheduleListener());
                 view.setRightPanel(panel);
@@ -133,13 +136,18 @@ public class HrController {
             JPanel panel = view.getCurrentPanel();
 
             if (panel instanceof PanelViewSchedule) {
-                PanelViewSchedule panelViewSchedule = (PanelViewSchedule) view.getCurrentPanel();
-                JTable scheduleTable = panelViewSchedule.getJTable();
-                panelViewSchedule.setVisible(false);
-                PanelInsertSchedule panelInsertSchedule=new PanelInsertSchedule(panelViewSchedule,schedules.get(0).getCnpEmployee());
-                panelInsertSchedule.addSaveScheduleListener(new SaveScheduleListener(null,schedules.get(0).getCnpEmployee()));
-                panelInsertSchedule.addCancelScheduleListener(new CancelScheduleListener());
-                view.setRightPanel(panelInsertSchedule);
+                if(model.viewRole(cnp)=="administrator" && (!account.isSuperAdmin())) {
+                    view.sendError("Nu aveti permisiunile necesare!");
+                }else {
+
+                    PanelViewSchedule panelViewSchedule = (PanelViewSchedule) view.getCurrentPanel();
+                    JTable scheduleTable = panelViewSchedule.getJTable();
+                    panelViewSchedule.setVisible(false);
+                    PanelInsertSchedule panelInsertSchedule = new PanelInsertSchedule(panelViewSchedule, schedules.get(0).getCnpEmployee());
+                    panelInsertSchedule.addSaveScheduleListener(new SaveScheduleListener(null, schedules.get(0).getCnpEmployee()));
+                    panelInsertSchedule.addCancelScheduleListener(new CancelScheduleListener());
+                    view.setRightPanel(panelInsertSchedule);
+                }
 
             }
         }
@@ -162,13 +170,17 @@ public class HrController {
                     return;
                 }
                 else {
-                    int row=scheduleTable.getSelectedRow();
-                    Schedule schedule= panelViewSchedule.getSchedules().get(row);
-                    panelViewSchedule.setVisible(false);
-                    PanelEditSchedule panelEditSchedule=new PanelEditSchedule(panelViewSchedule,schedule,account);
-                    panelEditSchedule.addSaveScheduleListener(new SaveScheduleListener(schedule,schedule.getCnpEmployee()));
-                    panelEditSchedule.addCancelScheduleListener(new CancelScheduleListener());
-                    view.setRightPanel(panelEditSchedule);
+                    if(model.viewRole(cnp)=="administrator" && (!account.isSuperAdmin())) {
+                        view.sendError("Nu aveti permisiunile necesare!");
+                    }else {
+                        int row = scheduleTable.getSelectedRow();
+                        Schedule schedule = panelViewSchedule.getSchedules().get(row);
+                        panelViewSchedule.setVisible(false);
+                        PanelEditSchedule panelEditSchedule = new PanelEditSchedule(panelViewSchedule, schedule, account);
+                        panelEditSchedule.addSaveScheduleListener(new SaveScheduleListener(schedule, schedule.getCnpEmployee()));
+                        panelEditSchedule.addCancelScheduleListener(new CancelScheduleListener());
+                        view.setRightPanel(panelEditSchedule);
+                    }
                 }
                 scheduleTable.clearSelection();
 
@@ -261,14 +273,21 @@ public class HrController {
                     view.sendError("Trebuie sa selectezi cel putin un rand.");
                     return;
                 }
-
+                int i=indexes.length-1;
                 for (int index : indexes) {
-                    model.deleteEmployeeSchedule(cnp,(Integer) scheduleTable.getValueAt(index, 0), (String) scheduleTable.getValueAt(index, 1), (Time) scheduleTable.getValueAt(index, 2),
-                            (Time)  scheduleTable.getValueAt(index, 3),1);
+                    if(model.viewRole(cnp)=="administrator" && (!account.isSuperAdmin())) {
+                        view.sendError("Nu aveti permisiunile necesare!");
+                        i--;
+                        continue;
+                    }else {
+                        model.deleteEmployeeSchedule(cnp, (Integer) scheduleTable.getValueAt(index, 0), (String) scheduleTable.getValueAt(index, 1), (Time) scheduleTable.getValueAt(index, 2),
+                                (Time) scheduleTable.getValueAt(index, 3), 1);
+                    }
                 }
 
                 panelViewSchedule.updateTable(model.viewSchedule(cnp));
-                view.sendSuccessMessage("Orarul selectat a fost sters.");
+                if(i>0)
+                    view.sendSuccessMessage("Orarul selectat a fost sters.");
             }
         }
     }
@@ -296,6 +315,7 @@ public class HrController {
             if (employeeTable.getSelectedRows().length != 1) {
                 view.sendError("Trebuie sa selectezi exact un angajat.");
             } else {
+
                 int row = employeeTable.getSelectedRow();
                 Employee employee = model.getEmployee((String) employeeTable.getValueAt(row, 3));
                 view.getViewv().setVisible(false);
@@ -347,13 +367,17 @@ public class HrController {
             JPanel panel = view.getCurrentPanel();
 
             if (panel instanceof PanelViewHoliday) {
-                PanelViewHoliday panelViewInsert = (PanelViewHoliday) view.getCurrentPanel();
-                JTable holidayTable = panelViewInsert.getJTable();
-                panelViewInsert.setVisible(false);
-                PanelInsertHoliday panelInsertHoliday=new PanelInsertHoliday(panelViewInsert,account);
-                panelInsertHoliday.addSaveHolidayListener(new SaveHolidayListener(null,holidays.get(0).getCnpEmployee()));
-                panelInsertHoliday.addCancelHolidayListener(new CancelScheduleListener());
-                view.setRightPanel(panelInsertHoliday);
+                if(model.viewRole(cnp)=="administrator" && (!account.isSuperAdmin())) {
+                    view.sendError("Nu aveti permisiunile necesare!");
+                }else {
+                    PanelViewHoliday panelViewInsert = (PanelViewHoliday) view.getCurrentPanel();
+                    JTable holidayTable = panelViewInsert.getJTable();
+                    panelViewInsert.setVisible(false);
+                    PanelInsertHoliday panelInsertHoliday = new PanelInsertHoliday(panelViewInsert, account);
+                    panelInsertHoliday.addSaveHolidayListener(new SaveHolidayListener(null, holidays.get(0).getCnpEmployee()));
+                    panelInsertHoliday.addCancelHolidayListener(new CancelScheduleListener());
+                    view.setRightPanel(panelInsertHoliday);
+                }
 
             }
         }
@@ -369,28 +393,32 @@ public class HrController {
             if (employeeTable.getSelectedRows().length != 1) {
                 view.sendError("Trebuie sa selectezi exact un angajat.");
             } else {
-                int row = employeeTable.getSelectedRow();
-                Employee employee = model.getEmployee((String) employeeTable.getValueAt(row, 3));
-                if(employee.getPosition().equals("Asistent Medical"))
-                    employee=model.getNurse(employee);
-                if(employee.getPosition().equals("Medic")) {
-                    employee = model.getDoctor(employee);
+                if(model.viewRole((String) employeeTable.getValueAt(employeeTable.getSelectedRow(), 3))=="administrator" && (!account.isSuperAdmin())) {
+                    view.sendError("Nu aveti permisiunile necesare!");
+                }else {
+                    int row = employeeTable.getSelectedRow();
+                    Employee employee = model.getEmployee((String) employeeTable.getValueAt(row, 3));
+                    if (employee.getPosition().equals("Asistent Medical"))
+                        employee = model.getNurse(employee);
+                    if (employee.getPosition().equals("Medic")) {
+                        employee = model.getDoctor(employee);
+                    }
+                    view.getViewv().setVisible(false);
+                    PanelEditEmployee panel = new PanelEditEmployee(employee, account,employee.getCnp(),model);
+                    if (employee.getPosition().equals("Medic")) {
+                        panel.updateTable(model.getSpecialities(employee.getCnp()));
+                        panel.updateTable2(model.getAccreditations(employee.getCnp()));
+                        panel.updateCbSpeciality(model.getSpecialities());
+                        panel.updateCbAcc(model.getAccreditations());
+                    }
+                    panel.addAccComboBoxListener(new AccListener(employee.getCnp()));
+                    panel.addSpecialityComboBoxListener(new SpecialityListener(employee.getCnp()));
+                    panel.deleteAccListener(new DeleteAccListener(employee.getCnp()));
+                    panel.deleteSpecListener(new DeleteSpecialityListener(employee.getCnp()));
+                    panel.addSaveButtonListener(new SaveButtonListenerHr(employee));
+                    panel.addCancelButtonListener(new CancelButtonListenerHr());
+                    view.setRightPanel(panel);
                 }
-                view.getViewv().setVisible(false);
-                PanelEditEmployee panel = new PanelEditEmployee(employee,account);
-                if(employee.getPosition().equals("Medic")) {
-                    panel.updateTable(model.getSpecialities(employee.getCnp()));
-                    panel.updateTable2(model.getAccreditations(employee.getCnp()));
-                    panel.updateCbSpeciality(model.getSpecialities());
-                    panel.updateCbAcc(model.getAccreditations());
-                }
-                panel.addAccComboBoxListener(new AccListener(employee.getCnp()));
-                panel.addSpecialityComboBoxListener(new SpecialityListener(employee.getCnp()));
-                panel.deleteAccListener(new DeleteAccListener(employee.getCnp()));
-                panel.deleteSpecListener(new DeleteSpecialityListener(employee.getCnp()));
-                panel.addSaveButtonListener(new SaveButtonListenerHr(employee));
-                panel.addCancelButtonListener(new CancelButtonListenerHr());
-                view.setRightPanel(panel);
             }
             employeeTable.clearSelection();
         }
@@ -405,18 +433,21 @@ public class HrController {
 
         @Override
         public void actionPerformed(ActionEvent e) {
-            PanelEditEmployee panel= (PanelEditEmployee) view.getCurrentPanel();
-            Map.Entry<Integer, String> s = panel.getIdAccreditation();
-            if(!panel.getAccDoc().containsKey(s.getKey())) {
-                view.sendError("Acreditarea nu exista.");
-                return;
+            if(model.viewRole(cnp)=="administrator" && (!account.isSuperAdmin())) {
+                view.sendError("Nu aveti permisiunile necesare!");
+            }else {
+                PanelEditEmployee panel = (PanelEditEmployee) view.getCurrentPanel();
+                Map.Entry<Integer, String> s = panel.getIdAccreditation();
+                if (!panel.getAccDoc().containsKey(s.getKey())) {
+                    view.sendError("Acreditarea nu exista.");
+                    return;
+                }
+                Map<Integer, String> spe = panel.getAccDoc();
+                spe.remove(s.getKey(), s.getValue());
+                panel.updateTable2(spe);
+                model.deleteAccreditation(cnp, s.getKey());
+                view.sendSuccessMessage("Acreditarea a fost stearsa.");
             }
-            Map<Integer,String> spe = panel.getAccDoc();
-            spe.remove(s.getKey(),s.getValue());
-            panel.updateTable2(spe);
-            model.deleteAccreditation(cnp,s.getKey());
-            view.sendSuccessMessage("Acreditarea a fost stearsa.");
-
         }
     }
 
@@ -427,17 +458,21 @@ public class HrController {
         DeleteSpecialityListener (String cnp){this.cnp=cnp;}
         @Override
         public void actionPerformed(ActionEvent e) {
-            PanelEditEmployee panel= (PanelEditEmployee) view.getCurrentPanel();
-            Map.Entry<Integer, String> s = panel.getIdSpeciality();
-            if(!(panel.getSpecialitiesDoc().stream().anyMatch(speciality -> speciality.getId() == s.getKey()))) {
-                view.sendError("Specializarea nu exista.");
-                return;
+            if(model.viewRole(cnp)=="administrator" && (!account.isSuperAdmin())) {
+                view.sendError("Nu aveti permisiunile necesare!");
+            }else {
+                PanelEditEmployee panel = (PanelEditEmployee) view.getCurrentPanel();
+                Map.Entry<Integer, String> s = panel.getIdSpeciality();
+                if (!(panel.getSpecialitiesDoc().stream().anyMatch(speciality -> speciality.getId() == s.getKey()))) {
+                    view.sendError("Specializarea nu exista.");
+                    return;
+                }
+                List<Speciality> spe = panel.getSpecialitiesDoc();
+                spe.removeIf(speciality -> speciality.getId() == s.getKey());
+                panel.updateTable(spe);
+                model.deleteSpeciality(cnp, s.getKey());
+                view.sendSuccessMessage("Specializarea a fost stearsa.");
             }
-            List<Speciality> spe = panel.getSpecialitiesDoc();
-            spe.removeIf(speciality -> speciality.getId()==s.getKey());
-            panel.updateTable(spe);
-            model.deleteSpeciality(cnp,s.getKey());
-            view.sendSuccessMessage("Specializarea a fost stearsa.");
         }
     }
 
@@ -448,17 +483,21 @@ public class HrController {
         SpecialityListener (String cnp){this.cnp=cnp;}
         @Override
         public void actionPerformed(ActionEvent e) {
-            PanelEditEmployee panel= (PanelEditEmployee) view.getCurrentPanel();
-            Map.Entry<Integer, String> s = panel.getIdSpeciality();
-            if(panel.getSpecialitiesDoc().stream().anyMatch(speciality -> speciality.getId() == s.getKey())) {
-                view.sendError("Specialitatea exista deja.");
-                return;
+            if(model.viewRole(cnp)=="administrator" && (!account.isSuperAdmin())) {
+                view.sendError("Nu aveti permisiunile necesare!");
+            }else {
+                PanelEditEmployee panel = (PanelEditEmployee) view.getCurrentPanel();
+                Map.Entry<Integer, String> s = panel.getIdSpeciality();
+                if (panel.getSpecialitiesDoc().stream().anyMatch(speciality -> speciality.getId() == s.getKey())) {
+                    view.sendError("Specialitatea exista deja.");
+                    return;
+                }
+                List<Speciality> spe = panel.getSpecialitiesDoc();
+                spe.add(new Speciality(s.getKey(), s.getValue(), (String) panel.getCbRank().getSelectedItem()));
+                panel.updateTable(spe);
+                model.insertSpeciality(cnp, s.getKey(), (String) panel.getCbRank().getSelectedItem());
+                view.sendSuccessMessage("Specializarea a fost adaugata.");
             }
-            List<Speciality> spe = panel.getSpecialitiesDoc();
-            spe.add(new Speciality(s.getKey(),s.getValue(), (String) panel.getCbRank().getSelectedItem()));
-            panel.updateTable(spe);
-            model.insertSpeciality(cnp,s.getKey(), (String) panel.getCbRank().getSelectedItem());
-            view.sendSuccessMessage("Specializarea a fost adaugata.");
         }
     }
 
@@ -469,17 +508,21 @@ public class HrController {
         AccListener (String cnp){this.cnp=cnp;}
         @Override
         public void actionPerformed(ActionEvent e) {
-            PanelEditEmployee panel= (PanelEditEmployee) view.getCurrentPanel();
-            Map.Entry<Integer, String> s = panel.getIdAccreditation();
-            if(panel.getAccDoc().containsKey(s.getKey())) {
-                view.sendError("Acreditarea exista deja.");
-                return;
+            if(model.viewRole(cnp)=="administrator" && (!account.isSuperAdmin())) {
+                view.sendError("Nu aveti permisiunile necesare!");
+            }else {
+                PanelEditEmployee panel = (PanelEditEmployee) view.getCurrentPanel();
+                Map.Entry<Integer, String> s = panel.getIdAccreditation();
+                if (panel.getAccDoc().containsKey(s.getKey())) {
+                    view.sendError("Acreditarea exista deja.");
+                    return;
+                }
+                Map<Integer, String> spe = panel.getAccDoc();
+                spe.put(s.getKey(), s.getValue());
+                panel.updateTable2(spe);
+                model.insertAccreditation(cnp, s.getKey());
+                view.sendSuccessMessage("Acreditarea a fost adaugata.");
             }
-            Map<Integer,String> spe = panel.getAccDoc();
-            spe.put(s.getKey(),s.getValue());
-            panel.updateTable2(spe);
-            model.insertAccreditation(cnp,s.getKey());
-            view.sendSuccessMessage("Acreditarea a fost adaugata.");
         }
     }
